Finish draft of init_process_volume_wf.
