+++ sdcflows/workflows/fit/medic.py
@@ -28,7 +28,17 @@ from nipype.interfaces.fsl import Split as FSLSplit
 from niworkflows.engine.workflows import LiterateWorkflow as Workflow
 
 from sdcflows.interfaces.fmap import PhaseMap2rads, ROMEO
-from sdcflows.utils.misc import calculate_diffs, medic_automask
+from sdcflows.utils.misc import (
+    calculate_diffs2,
+    calculate_fieldmap,
+    calculate_offset,
+    create_brain_mask,
+    global_mode_correction,
+    medic_automask,
+    modify_unwrapped_diff,
+    select_fieldmap,
+    subtract_offset,
+)
 
 INPUT_FIELDS = ("magnitude", "phase", "metadata")
 
@@ -232,29 +242,31 @@ def init_process_volume_wf(
         The magnitude EPI file that will be fed into MEDIC.
     phase : :obj:`str`
         The phase EPI file that will be fed into MEDIC.
+    mask : :obj:`str`
+        The brain mask that will be used to constrain the fieldmap estimation.
+        If ``automask`` is True, this mask will be modified.
+        Otherwise, it will be returned in the outputnode unmodified.
 
     Outputs
     -------
-    fmap : :obj:`str`
-        The path of the estimated fieldmap.
-    fmap_ref : :obj:`str`
-        The path of an unwarped conversion of files in ``in_data``.
-    fmap_mask : :obj:`str`
-        The path of mask corresponding to the ``fmap_ref`` output.
+    phase_unwrapped : :obj:`str`
+        Unwrapped phase in radians.
+    mask : :obj:`str`
+        Path to a brain mask that can be used to constrain the fieldmap estimation.
     """
     workflow = Workflow(name=name)
 
     inputnode = pe.Node(
-        niu.IdentityInterface(fields=["magnitude", "phase"]),
+        niu.IdentityInterface(fields=["magnitude", "phase", "mask"]),
         name="inputnode",
     )
     outputnode = pe.Node(
-        niu.IdentityInterface(fields=["fmap", "fmap_ref", "fmap_mask"]),
+        niu.IdentityInterface(fields=["phase_unwrapped", "mask"]),
         name="outputnode",
     )
 
     mask_buffer = pe.Node(
-        niu.IdentityInterface(fields=["mask_file", "masksum_file"]),
+        niu.IdentityInterface(fields=["mask"]),
         name="mask_buffer",
     )
     if automask:
@@ -263,10 +275,9 @@ def init_process_volume_wf(
         # but is noisy so we combine the two masks to get a better mask
 
         # Use ROMEO's voxel-quality command
-        voxqual = pe.MapNode(
+        voxqual = pe.Node(
             ROMEO(write_quality=True, echo_times=echo_times),
             name="voxqual",
-            iterfield=["in_file"],
         )
         workflow.connect([
             (inputnode, voxqual, [
@@ -280,7 +291,7 @@ def init_process_volume_wf(
         automask_medic = pe.Node(
             niu.Function(
                 input_names=["mag_file", "voxel_quality", "echo_times", "automask_dilation"],
-                output_names=["mask_file", "masksum_file"],
+                output_names=["mask", "masksum_file"],
                 function=medic_automask,
             ),
             name="automask_medic",
@@ -290,34 +301,40 @@ def init_process_volume_wf(
         workflow.connect([
             (inputnode, automask_medic, [("magnitude", "mag_file")]),
             (voxqual, automask_medic, [("quality_file", "voxel_quality")]),
-            (automask_medic, mask_buffer, [
-                ("mask_file", "mask_file"),
-                ("masksum_file", "masksum_file"),
-            ]),
+            (automask_medic, mask_buffer, [("mask", "mask")]),
         ])  # fmt:skip
     else:
-        mask_buffer.inputs.mask_file = "NULL"
-        mask_buffer.inputs.masksum_file = "NULL"
+        workflow.connect([(inputnode, mask_buffer, [("mask", "mask")])])
+
+    workflow.connect([(mask_buffer, outputnode, [("mask", "mask")])])
 
     # Do MCPC-3D-S algo to compute phase offset
-    create_diffs = pe.MapNode(
-        niu.Function(
-            input_names=["mag_file", "phase_file"],
-            output_names=["phase_offset", "unwrapped_diff", "phase_modified"],
-            function=calculate_diffs,
-        ),
-        iterfield=["mag_file", "phase_file"],
-        name="create_diffs",
-    )
+    mcpc_3d_s_wf = init_mcpc_3d_s_wf(wrap_limit=False, name="mcpc_3d_s_wf")
+    mcpc_3d_s_wf.inputs.inputnode.echo_times = echo_times
     workflow.connect([
-        (inputnode, create_diffs, [
-            ("magnitude", "mag_file"),
-            ("phase", "phase_file"),
+        (inputnode, mcpc_3d_s_wf, [
+            ("magnitude", "inputnode.magnitude"),
+            ("phase", "inputnode.phase"),
         ]),
+        (mask_buffer, mcpc_3d_s_wf, [("mask", "inputnode.mask")]),
     ])  # fmt:skip
 
-    # Unwrap phase data with ROMEO
-    unwrap_phase = pe.MapNode(
+    # remove offset from phase data
+    remove_offset = pe.Node(
+        niu.Function(
+            input_names=["phase", "offset"],
+            output_names=["phase_modified"],
+            function=subtract_offset,
+        ),
+        name="remove_offset",
+    )
+    workflow.connect([
+        (inputnode, remove_offset, [("phase", "phase")]),
+        (mcpc_3d_s_wf, remove_offset, [("outputnode.offset", "offset")]),
+    ])  # fmt:skip
+
+    # Unwrap the modified phase data with ROMEO
+    unwrap_phase = pe.Node(
         ROMEO(
             echo_times=echo_times,
             weights="romeo",
@@ -327,17 +344,31 @@ def init_process_volume_wf(
             correct_regions=False,
         ),
         name="unwrap_phase",
-        iterfield=["phase_file", "magnitude_file", "mask_file"],
     )
     workflow.connect([
         (inputnode, unwrap_phase, [("magnitude", "magnitude_file")]),
         (mask_buffer, unwrap_phase, [("mask_file", "mask_file")]),
-        (create_diffs, unwrap_phase, [("phase_modified", "phase_file")]),
+        (remove_offset, unwrap_phase, [("phase_modified", "phase_file")]),
     ])  # fmt:skip
 
     # Global mode correction
-
-    # Re-split the unwrapped phase data
+    global_mode_corr = pe.Node(
+        niu.Function(
+            input_names=["magnitude", "unwrapped", "mask", "echo_times"],
+            output_names=["unwrapped"],
+            function=global_mode_correction,
+        ),
+        name="global_mode_corr",
+    )
+    workflow.connect([
+        (inputnode, global_mode_corr, [
+            ("magnitude", "magnitude"),
+            ("echo_times", "echo_times"),
+        ]),
+        (unwrap_phase, global_mode_corr, [("out_file", "unwrapped")]),
+        (mask_buffer, global_mode_corr, [("mask", "mask")]),
+        (global_mode_corr, outputnode, [("unwrapped", "phase_unwrapped")]),
+    ])  # fmt:skip
 
     return workflow
 
@@ -558,187 +589,6 @@ def init_mcpc_3d_s_wf(wrap_limit, name):
     return workflow
 
 
-def calculate_diffs2(magnitude, phase):
-    """Calculate the magnitude and phase differences between two complex-valued images.
-
-    Parameters
-    ----------
-    magnitude : :obj:`str`
-        The path to the magnitude image (concatenated across echoes).
-    phase : :obj:`str`
-        The path to the phase image (concatenated across echoes).
-
-    Returns
-    -------
-    mag_diff_file : :obj:`str`
-        The path to the magnitude difference image.
-    phase_diff_file : :obj:`str`
-        The path to the phase difference image.
-    """
-    import os
-
-    import nibabel as nb
-    import numpy as np
-
-    mag_diff_file = os.path.abspath("magnitude_diff.nii.gz")
-    phase_diff_file = os.path.abspath("phase_diff.nii.gz")
-
-    magnitude_img = nb.load(magnitude)
-    phase_img = nb.load(phase)
-    magnitude_data = magnitude_img.get_fdata()
-    phase_data = phase_img.get_fdata()
-
-    signal_diff = (
-        magnitude_data[..., 0]
-        * magnitude_data[..., 1]
-        * np.exp(1j * (phase_data[..., 1] - phase_data[..., 0]))
-    )
-    mag_diff = np.abs(signal_diff)
-    phase_diff = np.angle(signal_diff)
-    mag_diff_img = nb.Nifti1Image(mag_diff, magnitude_img.affine, magnitude_img.header)
-    phase_diff_img = nb.Nifti1Image(phase_diff, phase_img.affine, phase_img.header)
-    mag_diff_img.to_filename(mag_diff_file)
-    phase_diff_img.to_filename(phase_diff_file)
-
-    return mag_diff_file, phase_diff_file
-
-
-def create_brain_mask(magnitude, extra_dilation):
-    """Create a quick brain mask for a single frame.
-
-    Parameters
-    ----------
-    magnitude : npt.NDArray[np.float32]
-        Magnitude data with the shortest echo time
-    extra_dilation : int
-        Number of extra dilations (or erosions if negative) to perform, by default 0
-
-    Returns
-    -------
-    npt.NDArray[np.bool_]
-        Mask of voxels to use for unwrapping
-    """
-    import os
-    from typing import cast
-
-    import nibabel as nb
-    import numpy as np
-    import numpy.typing as npt
-    from scipy.ndimage import (
-        binary_dilation,
-        binary_fill_holes,
-        binary_erosion,
-        generate_binary_structure,
-    )
-    from skimage.filters import threshold_otsu
-
-    from sdcflows.utils.misc import get_largest_connected_component
-
-    mag_img = nb.load(magnitude)
-    mag_data = mag_img.get_fdata()
-    mag_data = mag_data[..., 0]
-
-    mask_file = os.path.abspath("mask.nii.gz")
-
-    # create structuring element
-    strel = generate_binary_structure(3, 2)
-
-    # get the otsu threshold
-    threshold = threshold_otsu(mag_data)
-    mask_data = mag_data > threshold
-    mask_data = cast(npt.NDArray[np.float32], binary_fill_holes(mask_data, strel))
-
-    # erode mask
-    mask_data = cast(
-        npt.NDArray[np.bool_],
-        binary_erosion(mask_data, structure=strel, iterations=2, border_value=1),
-    )
-
-    # get largest connected component
-    mask_data = get_largest_connected_component(mask_data)
-
-    # dilate the mask
-    mask_data = binary_dilation(mask_data, structure=strel, iterations=2)
-
-    # extra dilation to get areas on the edge of the brain
-    if extra_dilation > 0:
-        mask_data = binary_dilation(mask_data, structure=strel, iterations=extra_dilation)
-    # if negative, erode instead
-    if extra_dilation < 0:
-        mask_data = binary_erosion(mask_data, structure=strel, iterations=abs(extra_dilation))
-
-    # since we can't have a completely empty mask, set all zeros to ones
-    # if the mask is all empty
-    if np.all(np.isclose(mask_data, 0)):
-        mask_data = np.ones(mask_data.shape)
-
-    mask_data = mask_data.astype(np.bool_).astype(np.uint8)
-    mask_img = nb.Nifti1Image(mask_data, mag_img.affine, mag_img.header)
-    mask_img.to_filename(mask_file)
-    return mask_file
-
-
-def calculate_offset(phase, unwrapped_diff, echo_times):
-    """Calculate the phase offset between two echoes.
-
-    Parameters
-    ----------
-    phase : :obj:`str`
-        The path to the phase image (concatenated across echoes).
-    unwrapped_diff : :obj:`str`
-        The path to the unwrapped phase difference image from the first two echoes.
-    echo_times : :obj:`list` of :obj:`float`
-        The echo times.
-
-    Returns
-    -------
-    offset : :obj:`str`
-        The path to the phase offset image.
-    """
-    import os
-
-    import nibabel as nb
-    import numpy as np
-
-    offset_file = os.path.abspath("offset.nii.gz")
-
-    phase_img = nb.load(phase)
-    unwrapped_diff_img = nb.load(unwrapped_diff)
-    phase_data = phase_img.get_fdata()
-    unwrapped_diff_data = unwrapped_diff_img.get_fdata()
-
-    proposed_offset = np.angle(
-        np.exp(
-            1j
-            * (
-                phase_data[..., 0]
-                - ((echo_times[0] * unwrapped_diff_data) / (echo_times[1] - echo_times[0]))
-            )
-        )
-    )
-    proposed_offset_img = nb.Nifti1Image(proposed_offset, phase_img.affine, phase_img.header)
-    proposed_offset_img.to_filename(offset_file)
-    return offset_file
-
-
-def subtract_offset(phase, offset):
-    import os
-
-    import nibabel as nb
-    import numpy as np
-
-    updated_phase_file = os.path.abspath("updated_phase.nii.gz")
-
-    phase_img = nb.load(phase)
-    offset_img = nb.load(offset)
-    phase_data = phase_img.get_fdata()
-    offset_data = offset_img.get_fdata()
-    updated_phase = phase_data - offset_data[..., np.newaxis]
-    updated_phase_img = nb.Nifti1Image(updated_phase, phase_img.affine, phase_img.header)
-    updated_phase_img.to_filename(updated_phase_file)
-    return updated_phase_file
-
-
 def init_dual_echo_wf(name="dual_echo_wf"):
     workflow = Workflow(name=name)
 
@@ -796,163 +646,3 @@ def init_dual_echo_wf(name="dual_echo_wf"):
     ])  # fmt:skip
 
     return workflow
-
-
-def calculate_fieldmap(unwrapped_phase, echo_times):
-    import os
-
-    import nibabel as nb
-    import numpy as np
-
-    fieldmap_file = os.path.abspath("fieldmap.nii.gz")
-
-    unwrapped_phase_img = nb.load(unwrapped_phase)
-    unwrapped_phase_data = unwrapped_phase_img.get_fdata()
-
-    phase_diff = unwrapped_phase_data[..., 1] - unwrapped_phase_data[..., 0]
-    fieldmap = (1000 / (2 * np.pi)) * phase_diff / (echo_times[1] - echo_times[0])
-    fieldmap_img = nb.Nifti1Image(
-        fieldmap,
-        unwrapped_phase_img.affine,
-        unwrapped_phase_img.header,
-    )
-    fieldmap_img.to_filename(fieldmap_file)
-
-    return fieldmap_file
-
-
-def modify_unwrapped_diff(phase, unwrapped_diff, echo_times):
-    import os
-
-    import nibabel as nb
-    import numpy as np
-
-    updated_phase_file = os.path.abspath("updated_phase.nii.gz")
-
-    phase_img = nb.load(phase)
-    unwrapped_diff_img = nb.load(unwrapped_diff)
-    phase_data = phase_img.get_fdata()
-    unwrapped_diff_data = unwrapped_diff_img.get_fdata()
-
-    new_proposed_offset = np.angle(
-        np.exp(
-            1j
-            * (
-                phase_data[..., 0]
-                - (
-                    (echo_times[0] * (unwrapped_diff_data + 2 * np.pi))
-                    / (echo_times[1] - echo_times[0])
-                )
-            )
-        )
-    )
-    new_proposed_phase = phase_data - new_proposed_offset[..., np.newaxis]
-    new_proposed_phase_img = nb.Nifti1Image(
-        new_proposed_phase,
-        phase_img.affine,
-        phase_img.header,
-    )
-    new_proposed_phase_img.to_filename(updated_phase_file)
-
-    return updated_phase_file
-
-
-def select_fieldmap(
-    original_fieldmap,
-    original_unwrapped_phase,
-    original_offset,
-    modified_fieldmap,
-    modified_unwrapped_phase,
-    unwrapped_diff,
-    voxel_mask,
-    echo_times,
-    wrap_limit,
-):
-    import os
-
-    import nibabel as nb
-    import numpy as np
-
-    FMAP_PROPORTION_HEURISTIC = 0.25
-    FMAP_AMBIGUIOUS_HEURISTIC = 0.5
-
-    new_unwrapped_diff_file = os.path.abspath("new_unwrapped_diff.nii.gz")
-
-    orig_fmap_img = nb.load(original_fieldmap)
-    orig_unwrapped_phase_img = nb.load(original_unwrapped_phase)
-    orig_offset_img = nb.load(original_offset)
-    mod_fmap_img = nb.load(modified_fieldmap)
-    mod_unwrapped_phase_img = nb.load(modified_unwrapped_phase)
-    unwrapped_diff_img = nb.load(unwrapped_diff)
-    voxel_mask_img = nb.load(voxel_mask)
-
-    orig_fmap_data = orig_fmap_img.get_fdata()
-    orig_unwrapped_phase_data = orig_unwrapped_phase_img.get_fdata()
-    orig_offset_data = orig_offset_img.get_fdata()
-    mod_fmap_data = mod_fmap_img.get_fdata()
-    mod_unwrapped_phase_data = mod_unwrapped_phase_img.get_fdata()
-    unwrapped_diff_data = unwrapped_diff_img.get_fdata()
-    voxel_mask_data = voxel_mask_img.get_fdata()
-
-    masked_orig_fmap = orig_fmap_data[voxel_mask_data]
-    if masked_orig_fmap.mean() < -10:
-        # check if the proposed fieldmap is below -10
-        unwrapped_diff_data += 2 * np.pi
-
-    elif masked_orig_fmap.mean() < 0 and not wrap_limit:
-        # check if the proposed fieldmap is between -10 and 0
-        # look at proportion of voxels that are positive
-        voxel_prop = np.count_nonzero(masked_orig_fmap > 0) / masked_orig_fmap.shape[0]
-
-        # if the proportion of positive voxels is less than 0.25, then add 2pi
-        if voxel_prop < FMAP_PROPORTION_HEURISTIC:
-            unwrapped_diff_data += 2 * np.pi
-        elif voxel_prop < FMAP_AMBIGUIOUS_HEURISTIC:
-            # compute mean of phase offset
-            mean_phase_offset = orig_offset_data[voxel_mask_data].mean()
-            if mean_phase_offset < -1:
-                phase_fits = np.concatenate(
-                    (
-                        np.zeros((*orig_unwrapped_phase_data.shape[:-1], 1)),
-                        orig_unwrapped_phase_data,
-                    ),
-                    axis=-1,
-                )
-                _, residuals_1, _, _, _ = np.polyfit(
-                    echo_times,
-                    phase_fits[voxel_mask_data, :].T,
-                    1,
-                    full=True,
-                )
-
-                masked_mod_fmap = mod_fmap_data[voxel_mask_data]
-                # fit linear model to the proposed phase
-                new_phase_fits = np.concatenate(
-                    (
-                        np.zeros((*mod_unwrapped_phase_data.shape[:-1], 1)),
-                        mod_unwrapped_phase_data,
-                    ),
-                    axis=-1,
-                )
-                _, residuals_2, _, _, _ = np.polyfit(
-                    echo_times,
-                    new_phase_fits[voxel_mask_data, :].T,
-                    1,
-                    full=True,
-                )
-
-                if (
-                    np.isclose(residuals_1.mean(), residuals_2.mean(), atol=1e-3, rtol=1e-3)
-                    and masked_mod_fmap.mean() > 0
-                ):
-                    unwrapped_diff_data += 2 * np.pi
-                else:
-                    unwrapped_diff_data -= 2 * np.pi
-
-    new_unwrapped_diff_img = nb.Nifti1Image(
-        unwrapped_diff_data,
-        unwrapped_diff_img.affine,
-        unwrapped_diff_img.header,
-    )
-    new_unwrapped_diff_img.to_filename(new_unwrapped_diff_file)
-    return new_unwrapped_diff_file
